- reading of 16-bit "first/last" values in RLE shapes made independent   of endianess - small fixes to make it compile with gcc under Linux
git-svn-id: svn://svn.code.sf.net/p/marathon/code/branches/LP_fileio@471 a75d40c9-0969-4759-85e5-97cb871b7ed4

+++ Source_Files/Misc/shapes.cpp
@@ -537,7 +537,7 @@ static boolean load_collection(
 			assert(Definition.bitmap_offset_table_offset >= 0);
 			assert(Definition.bitmap_offset_table_offset +
 				Definition.bitmap_count*sizeof(int32) <= length);
-			long *BtmpOffsets = new long[Definition.bitmap_count + 1];
+			int32 *BtmpOffsets = new int32[Definition.bitmap_count + 1];
 			S = collection + Definition.bitmap_offset_table_offset;
 			StreamToList(S,BtmpOffsets,Definition.bitmap_count);
 			BtmpOffsets[Definition.bitmap_count] = length;
